Driving Works Perfectly
Driving bug has been completely fixed! And voltage, speed, and position
mode work correctly. Current mode has not been tested enough to know
exactly how it works, but it is functional at the moment (depending on
the input value you give it).

This is a safe version to revert back to